Create PowerSploitUpdate2021
I modified the original version of Get-GPPPassword because if it didn't find any cPassword instances then it returned nothing.  I wanted it to confirm that it found nothing.  I also added a way to test it by changing the text it searches for in the xml files from "cPassword" to "Group" just as a test.